Add support for local clickhouse in docker (#48)
Caveats / Deficiencies

This is not a perfect match for the server clickhouse config since we're
using MergeTree() as the engine. It seemed like it was going to be much
more complex to exactly match production, but I may have missed something
obvious.

The create tables sql only seems to be run on a completely fresh install.